refactor: reordering network-model files
