+++ app/src/main/java/com/arkan/terbangin/data/datasource/auth/login/LoginApiDataSource.kt
@@ -1,6 +1,6 @@
 package com.arkan.terbangin.data.datasource.auth.login
 
-import com.arkan.terbangin.data.source.network.model.login.LoginResponse
+import com.arkan.terbangin.data.source.network.model.auth.login.LoginResponse
 import com.arkan.terbangin.data.source.network.services.TerbanginApiServices
 import okhttp3.RequestBody
