+++ app/src/main/java/com/arkan/terbangin/data/datasource/auth/login/LoginDataSource.kt
@@ -1,6 +1,6 @@
 package com.arkan.terbangin.data.datasource.auth.login
 
-import com.arkan.terbangin.data.source.network.model.login.LoginResponse
+import com.arkan.terbangin.data.source.network.model.auth.login.LoginResponse
 import okhttp3.RequestBody
 
 interface LoginDataSource {
