+++ app/src/main/java/com/arkan/terbangin/data/datasource/auth/resetpassword/ResetPasswordApiDataSource.kt
@@ -1,6 +1,6 @@
 package com.arkan.terbangin.data.datasource.auth.resetpassword
 
-import com.arkan.terbangin.data.source.network.model.resetpassword.ResetPasswordResponse
+import com.arkan.terbangin.data.source.network.model.auth.resetpassword.ResetPasswordResponse
 import com.arkan.terbangin.data.source.network.services.TerbanginApiServices
 import okhttp3.RequestBody
