+++ app/src/main/java/com/arkan/terbangin/data/repository/auth/LoginRepository.kt
@@ -3,7 +3,7 @@ package com.arkan.terbangin.data.repository.auth
 import android.util.Log
 import com.arkan.terbangin.data.datasource.auth.login.LoginDataSource
 import com.arkan.terbangin.data.datasource.preference.PreferenceDataSource
-import com.arkan.terbangin.data.source.network.model.login.LoginResponse
+import com.arkan.terbangin.data.source.network.model.auth.login.LoginResponse
 import com.arkan.terbangin.utils.ResultWrapper
 import com.arkan.terbangin.utils.proceedFlow
 import kotlinx.coroutines.flow.Flow
