+++ app/src/main/java/com/arkan/terbangin/data/source/network/services/TerbanginApiServices.kt
@@ -1,11 +1,11 @@
 package com.arkan.terbangin.data.source.network.services
 
 import com.arkan.terbangin.BuildConfig
-import com.arkan.terbangin.data.source.network.model.login.LoginResponse
-import com.arkan.terbangin.data.source.network.model.otp.request_otp.RequestOTPResponse
-import com.arkan.terbangin.data.source.network.model.otp.verify_otp.VerifyOTPResponse
-import com.arkan.terbangin.data.source.network.model.register.RegisterResponse
-import com.arkan.terbangin.data.source.network.model.resetpassword.ResetPasswordResponse
+import com.arkan.terbangin.data.source.network.model.auth.login.LoginResponse
+import com.arkan.terbangin.data.source.network.model.auth.otp.request_otp.RequestOTPResponse
+import com.arkan.terbangin.data.source.network.model.auth.otp.verify_otp.VerifyOTPResponse
+import com.arkan.terbangin.data.source.network.model.auth.register.RegisterResponse
+import com.arkan.terbangin.data.source.network.model.auth.resetpassword.ResetPasswordResponse
 import com.arkan.terbangin.data.source.pref.UserPreference
 import okhttp3.OkHttpClient
 import okhttp3.RequestBody
